+++ src/main/java/com/javarush/domain/Staff.java
@@ -16,7 +16,7 @@ public class Staff {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "staff_id")
-    private Integer id;
+    private Byte id;
 
     @Column (name = "first_name")
     private String firstName;
